There's a bug which means you can't click the side menu 'basic dwelling data' link and then add floors.
If you click 'basic dwellins data', clicking the 'add floors button' has no effect. We discovered the side menu link doesn't load `context.js` (and so doesn't attach the event to the button). This is different behaviour to the other side menu links.

We've worked around this by populating the floor data as soon as you open the assessment, as this information is on the 'front' page.